Extract `getAssets`
